Add a cheeky CRC on the HS token.
This makes sure we're all singing from the same hymn sheet when it comes to
user prefixes, alias prefixes, etc. A mismatch would cause horrendously weird
and hard to debug errors, so we append a CRC to ensure that inbound HS requests
and loaded ASes are referring to the same config.yaml.