Gut CAS auth from the AS.
The problem with it is that this leaks events from 'private' (behind CAS)
rooms over federation if a user on @foo:bar auths correctly using CAS. This
is deemed a big enough problem to shuffle the architecture/checks to prevent
federation in this case.

Decided that it would be better to do CAS auth on the HS and then restrict
rooms by user_id domain part in order to prevent private channels being
federated.