BOTS-29 : Fix duplicate IRC nicks for some users.
When the AS receives incoming Matrix messages, it connects an IRC user if one
isn't connected already then sends the message. If you received 2+ messages in
quick succession, the 2nd message would incorrectly trigger another connect if
the on connected callback from the first message hadn't fired yet. We now store
the IRC connection as soon as we know it doesn't exist in order to fix this
race. When sending messages, we now wait on a connection promise since now
cached clients may not be connected yet so you can't send messages. Added a
regression test for this race.