BOTS-41 : Fix echo when nicks clash.
This was caused by clobbering the VirtualIrcUser entry with a nick which was
not verified by the IRC server (pre-connected). Also modify the config
validator to return a POJO rather than a Config object as only POJOs are deep
copied by the extend library. This meant that test configuration in one file
was leaking through to other tests in other files even with extend(). Finally,
tweak client cycling to take pre-connected clients into account.