BOTS-39 : Always create matrix users when queried.
In cases where we mirror join/parts, it's possible to get onUserQuery events
for IRC users who are not on /whois. Previously, this would result in the user
not being registered. We now register the user to avoid the HS sending lots
of onUserQuery requests for unregistered user IDs. This does mean that typoed
IRC user IDs will be provisioned, but the footprint on the HS/AS is tiny, so
is probably okay.