Major reorganisation of IRC client connection code
Move connection related logic to a new file client-connection.js and more
clearly separate concerns with reconnecting dodgy connections and making
the connection in the first place. A new class, IrcasClient, can connect
a client but will just try once, which may fail due to any number of
reasons (throttling; no connection; timeout; etc).

The create() method of client-connection.js is tasked with attempting
connections a number of times before giving up. This is all abstracted
behind a promise interface which just resolves with a working client
or rejects with an error message. This allows us to completely blow away
the node-irc client object in the event of a connection problem.

More work needs to be done wrt managing event listeners on dead connections
and managing reconnections.