Make VirtualIrcUser the canonical place for IRC client interactions.
Previously, the bot would directly access the node-irc client, which we
are now wrapping. This actually ends up reducing a lot of duplication as
the virtual clients and the bot tend to do the same thing. Combining them
into VirtualIrcUser means any connection logic (reconnects in particular)
will benefit both bot and client and keeps things DRY.

IrcServer.connectAs() needs to be folded into VirtualIrcUser, and then
reconnect logic needs to be added to VirtualIrcUser.