Time out join requests after JOIN_TIMEOUT_MS and try again
It's possible for some join requests to get dropped by the ircd, seen on
freenode. To handle this, wait for 15s after sending the JOIN and if we
haven't got a response to join that channel (and we haven't got an ERROR
to say err_bannedfromchan/err_toomanychannels/etc) then try again. Fix
racy tests.