Fix BOTS-139 (prioritise unique rooms first on connect)
This limits the impact of downtime between restarts as this favours connecting
1 matrix user to each IRC channel before connecting others. This is important
to keep the flow of IRC->Matrix traffic going.

This system can be extended if the injected join event marks if it is a frontier
event. If it is marked, the bridge can use this info to forcibly join that user
to the channel even if mirroring membership lists is disabled iff there is no
bot (else you get no traffic at all!)