Use coroutines in irc-to-matrix callbacks
This greatly simplifies the code and has caught a few bugs in the process.
Also, inject a Request object into the callback rather than awkwardly creating
it inside the callback then manage the sucFn/errFn logic. Just have the callback
resolve/reject appropriately which affects the outcome of the Request. This
feels significantly better from before.