Tests: Remove registration v1 hack. Unbreak and clarify admin room test
The admin room test would test for "join if known room ID" using a BRIDGED
room and clobber it with an ADMIN room - we don't really support this.

This came to light when the UNIQUE indexes were added: The UPSERT would check
for an existing admin room, fail to find one, then insert a new admin room which
would then conflict with the existing bridged room.