Modify when connection notices are sent (Fixes #32 and #107)
Previously, connection notices were sent when the TCP connection was made, when
what we really wanted is to send the connection notice after receiving
rpl_welcome. This commit makes notices be sent after rpl_welcome.

This commit also removes the listener for 'registered' events, which was there
to catch nick mismatches on rpl_welcome. Where it was placed meant that it
would never fire, as the listener was added *AFTER* ConnectionInstance.connect
returned a resolved promise (the yield). That promise only resolves when
the callback to node-irc's `client.connect(callback)` fires, which only fires
*AFTER* the rpl_welcome and the 'registered' event.

This broke the test for "using the nick in rpl_welcome", but the test was wrong.
The test was erroneously firing a *SECOND* 'registered' event (effectively
another rpl_welcome command) after the connection had been established, which
then changed the nick. The test now reflects reality.