Fix #25 by kicking Matrix users when their IRC connections get certain error codes
These include bans/bad keys/invite only/full channels.

We cannot always prevent a Matrix user from joining a Matrix room. However, we
can check to see if we get any errors when trying to join the corresponding
IRC channel and then kick the user if the error means they should not be in
the room.