Removed imediate reconnection
Imediately trying to reconnect defeats the purpose of the scheduler.
Even if the interval was reset in the case of the 'first' reconnection,
this could allow any number of imediate connections. That is unless the
procFn of a queue was forced to last at least as long as the interval.