Ask IRC Op for Bridging
Link provisioning now requires that the nick of a valid IRC operator
is given with a requested link. The bridge sends a message to the op
asking to respond 'yes' if they would like to authorise a given
bridging.

Determining whether someone is an op uses whois, which required tweaking
on the BridgedClient to use the operator field that was so far unused.

m.room.bridging room state is used in the room in the new mapping to
signal whether the status of the bridging is 'pending', 'success', or
'failure'.
 = {
    content: {
        bridger: 'nick',
        status: 'pending'
    }
}