Fix bugs following updated tests
The logic surrounding power levels was previously bugged
to allow users with 0 power to have infinite power... This
has been fixed and I should remember not to use '||' when
the user's power could be 0 (which is true by default).

Also, protecting dolink with a try catch that prevents
m.room.bridging = success even upon failure.