Implement /_matrix/provision/queryLink
This can be used to do a 'dummy' link request to check for errors with parameters. Currently, no further detail on the validation of parameters is given better than 'Parameters malformed'. This is an issue with ConfigValidator component of matrix-appservice-bridge.

Results from this endpoint are given as a JSON blob, for example :

```JavaScript
    {
	parameter_info: {
	    remote_room_server: {},
	    remote_room_channel: {},
	    matrix_room_id: {},
	    op_nick: {error : 'User provided not an op in this channel'},
	    user_id: {}
	},
	// General info and errors
	infos: [],
	errors: [],
	// List of operators. Only defined if retrieved.
	operators: []
    }
```