Prevent loop routing when provisioning new mappings
The provisioner will fail to request for a mapping if the m.room.bridging status in a room for the same channel (i.e. with the same state key) is not = 'failure'. Meaning, if it is either pending or a success, the request will fail (unless the bridging requestor is the same as the user_id in the m.room.bridging).

Maybe the check for user_id should really be with the sender, meaning the bridge checks that its own matrix bot was the one to send the bridging. But currently, when you get an event, you don't get it's original sender, so we may have to add extra state to the content of the m.room.bridging indicating the matrix id of the bridge that set it.