Change queryLink to not validate everything
queryLink is now an endpoint that should be used to get information about a potential link as opposed to validating a hypothetical /link call. For now, this information is restricted to the operators in the IRC channel. If any errors occur, this results in an error being thrown, so there are no longer multiple 'errors' being returned.