Handle cases where bridge was restarted
If the bridge is restarted, _pendingRequests is wiped, so when a request is sent the bridge checks the bridging state to see if it had set the state previously. If it has and it is pending, it continues with the request, effectively reinstating the missing entry from _pendingRequests. If it has and it is success, it checks the database to see if a mapping exists (as in the link was previously successful) and if there is no such mapping, it resets the state to failure and then continues with the request.