Improved admin room details
Mostly to stop confusion regarding it's use/behaviour.

Upon joining a network, the bot will invite the user to a room called 'IRC Admin Room', if the admin room does not already exist. The message sent upon connecting to a network now looks like:

> You've been connected to the IRC network 'localhost' as Lukeb[m]. This room can be used to send commands directly through your personal connection to connected IRC networks. Type !help for more information.

Sending '!help' will now give:
```
This is an IRC admin room for sending commands directly to IRC.
Commands can be sent all as one message, but these will not reply with any information:
	irc.server
	(JOIN [arg0 [arg1 [...]]])
	irc.server2
	(KICK [arg0 [arg1 [...]]])
	(COMMAND [arg0 [arg1 [...]]])
Or if feedback is desired, the following commands can be used:
	!join irc.example.com #channel [key]
	!nick irc.example.com DesiredNick
	!whois nick
```