Improve admin room messages
When a client is connected to the IRC network, they now only see a minimal message in the IRC admin room, "You've been connected to _ as _". When the admin room is first created (only if one doesn't already exist) a message is sent to the admin room saying, "You've joined a Matrix room which is bridged to the IRC network '...', where you are now connected as '...'. This room shows any errors or status messags from IRC, as well as letting you control the connection. Type !help for more information);".