Remove `disabled` property on `BridgedClient`
It makes little semantic sense to cripple most methods of an instance like this.
The checks for "should I call this function or not" should be the responsibility
of the caller. This bodge was added in a long time ago to make it faster to
deploy a no-bot bridge. This has now been done properly, where the bridge
performs `server.isBotEnabled()` checks at key points (joining channels) rather
than blindly disabling huge wodges of functionality of a `BridgedClient`. This
is important because `Provisioning` *requires* a working bot client and not a
crippled one.