Use the sync version of unlink to try to resolve spurious Travis failures
Travis can fail random tests with:

  Error: ENOENT: no such file or directory, rename 'spec-db/users.db~' -> 'spec-db/users.db'
    at Error (native)

This isn't due to double `unlink()` calls (else the operation would be "unlink"
and not "rename"). We don't ever call `fs.rename`. It looks like Travis is
doing something Weird, or maybe a later version of Node is? A quick look through
the Node source didn't show anything interesting. We don't ever use the tilde
suffix either.

The error is **impossible** to catch because it's on a different turn of the
event loop and we didn't call `rename()`, so we can't make it not throw by
adding an `errback`, nor can we `try/catch` because it's async.

Swap to using `unlinkSync` in the vain hope that it will mean that `rename`
will also be called syncly and therefore can be `try/catch`d.