Do not cache stale clients. Always keep the bot client pool up-to-date
Turns out the `ClientPool` would not update the cache when the bot disconnected
then reconnected, resulting in a stale `BridgedClient` being returned when
calling `clientPool.getBot(server)`. This is now fixed.

In addition, do not have another layer of caching at the `Provisioner`, and
instead rely on `ClientPool` to not be broken. We need to remove this cache
else it too will get stale clients on disconnects.