Cache the visibilty of a room last sent to mx
This is so that the HS is not hammered on startup when a lot of RPL_CHANMODEIS could be received as clents join channels.

The problem with this is then what happens when two channels are mapped to one room and just one of the rooms is +s or +i/k?