Add membershipLists.floodDelayMs to config.yaml
This replaces `JOIN_WAIT_TIME_MS` which was hard-coded to 10s. This means we
can sync membership lists from M->I much more rapidly if we are allowed to
(e.g. by being on IPv6).

Also tweak how we invoked `joinNextUser()`. It appears that using `finally()`
is NOT enough to make bluebird not whine about uncaught exceptions, resulting
in `TimeoutErrors` in the logs. We now explicitly `then()` the Promise and just
do the same no matter what happens (which is semantically the same as `finally`).
This removes all those `TimeoutErrors` in the logs \o/