Stop the freenode bridge from spinning for minutes by fixing #242
The spin was caused by the bridge processing the `/initialSync` JSON response.
This is a huge wodge of JSON which took **minutes** to process into a JSON
object. This completely wedged the bridge for that length of time
(no logging, no ping/pongs, no messages being sent/recv).

This fixes it by never calling `/initialSync`, and instead always using
`/$room_id/state` even if there are >1 mapped Matrix rooms for a given IRC channel.