Enable simple presence syncing
When a user joins a channel, 'online' will be sent as their presence to the server. When a user parts and a 'quit' is indicated (i.e. the user quit), the bridge will set their presence to 'offline' and wait 15mins before causing the corresponding virtual matrix user to leave rooms. If the user joins a channel in that time, their presence will be set to 'online' and the corresponding virtual matrix user will not leave rooms.