Spelling and also fixed racing with net split detection
Now split detection is done in a way that doesn't race with other splits. Before, if multiple debounceQuit calls were made in quick succession, this would wedge the setInterval trying to do statistics on the number of quits occuring. This has been prevented by not using setInterval, but updating statistics every time debounceQuit is called in a concurrency-safe way.

The same stats are used to determine a net split, the threshold is > 5Hz of QUITs received. Less than that, and the leave will be bridged after a small 100ms delay to allow for other QUITs to be received before checking the current rate.