Store passwords in user.db user configs
The passwords are encrypted and decrypted using a PEM key stored in a file specified in the config file (See passwordEncryptionKeyPath).

The passwords are then used in place of the server password set in the config file.

This has been tested to work with freenode, and as expected, when !nick and !storepass are issued prior to the bridge restarting (or the IRC client reconnecting), NickServ asks foridentification (if the nick is registered) but then imediately resolves (having received PASS with the recently decrypted password).