Try to kick bridged mx users on !quit
When a !quit command is issued, make the bridge mx bot kick all instances of the bridged mx user from all bridged rooms. This currently excludes PM rooms and admin rooms and any rooms in which the user has an equal or greater power than the bot.

If users cannot be kicked from a room, log it but continue to disconnect the user from IRC.