Leave matrix users if their associated IRC nick is no longer in a bridged channel
On startup, matrix member lists are stored in memory for later reference. They may not be entirely accurate long after startup, but it won't stop the majority of users being cleaned up.

The ClientPool listens for any "irc-names" events being emitted from BridgedClients. This event is triggered whenever an IRC client receives NAMES. The ClientPool calls updateIrcMemberList on MemberListSyncer. This function leaves any matrix users that do not appear to be in the IRC member list of the channel whose mlist has just been updated from all bridged rooms of that channel.