Listen for err_yourebannedcreep and wait 6hrs before reconnecting
When `err_yourebannedcreep` is received, 2 events fire: `error` and `raw`. It
looks like `error` consistently happens before `raw`, which would lead to
`disconnect(irc_error)` which then flips the connection state to `dead`. When
the `raw` event fires, it tries to fail with `disconnect(banned)` but because
we've already flipped the connection state to `dead`, this does nothing.

As a consequence, banned connections would retry much more frequently than they
should.