PublicitySyncer: Use networkId instead of server domain
Instead of keying of the server domain for publicity syncing, use a network ID which should be unique amongst the servers being bridged to allow the server domain being changed without public room lists being affected. This networkId could be used in future as a way to key servers on the bridge, although because they are defined as strings they are not guaranteed to be unique.