Do not treat invites for unknown user IDs as errors
The bridge will get invites *it has sent* mirrored back to itself.
These were previously treated as FAILED because we didn't explicitly
handle this case. We now handle this case, and no longer have a bulk
try/catch around converting the invitee to an IRC user. This was part
of the problem as it was impossible to know if the error was due to
no mapping existing or a `TypeError` or equiv.