Let the bridge appear in the thirdparty/protocols API
In order to be listed in the thirdparty/protocols API, the
resigtration file must set the protocols setting.  Optionally an icon
can be set in the bridge configuration, that is used in the
thirdparty/protocols API accodingly. This makes the bridged network
to show up in the network pull-down list in the riot room directory,
for instance.