Make ident reliable by delaying negative response
The IRC server typically sends the ident request as soon as the TCP
connection is established. However, since we only add the ident mapping
after we get the established connection socket, the response is
extremely racy and often fails to find the mapping.

This commit follows [the scheme implemented by Quassel][1] and delays
giving a negative response while there are still connections in the
process of being opened and are older than the request. This should make
the response reliable.

[1]: https://github.com/quassel/quassel/pull/356