Avoid duplicating virtual users when using !nick
Currently when using !nick, the new nick is not yet mapped to the
Matrix user when the new name is received from the server by other
IRC clients, which leads to a new user joining the room if
incremental sync is configured.

Instead, this patch introduces events to add and remove temporary
mappings from nicks to matrix users to handle nick changes. Just
before the nick is changed, the new nick mapped to the Matrix
user, then the mapping is either removed if something wrong
happened or kept if the nick change was effective.