Fix uploadTextFile call to uploadContent
The call was changed in
https://github.com/matrix-org/matrix-appservice-irc/pull/845 to
instead of getting the raw response back, to get just the content
uri back. Due to wrong type definition of the uploadContent
method, the change actually made the raw response still get back
due to legacy defaults in uploadContent.

Refactor the call to pass the buffer as first argument and opts
as second. This has been preferred in matrix-js-sdk since 2016.

Refs: #887
Signed-off-by: Jason Robinson <jasonr@matrix.org>