Disable !join when dynamic channels are off
AFAICT this may lead to private, statically configured channels getting
exposed to the public, as `hasInviteRooms` is `false` when
`dynamicChannels.enabled` is off, thus the user is considered always
"authorised to join channels".

I haven't tested if this is actually the case; I just sleep better when
the `!join` command is immediately denied when `dynamicChannels.enabled`
is off.

Signed-off-by: Jan Alexander Steffens (heftig) <jan.steffens@gmail.com>