Reconnect to the correct domain on passsword changes
The appservice triggers a reconnect when handling "!storepass" commands, but
rather than reconnecting to the domain for which the password was changed the
appservice just always reconnects to the first domain.  This causes an issue
for my setup, which has multiple domains in a singe appservice.

I've tested this by running it on my instance, which is connected to two
networks, and changing my password.

Signed-off-by: Palmer Dabbelt <palmerdabbelt@google.com>