Fix matrix name change not propagating to irc
When we receive a matrix join event, we compare
bridgedClient.displayName with the new display name to see if we need to
update the irc client's nick. However, we didn't set
bridgedClient.displayName to the new nick afterwards. This means if the
matrix user changes their display name back to the original, the change
is not propagated to irc.