config.schema.yml: fix permissions schema
The key reflects a key, value mapping from mxid to "admin", and
additionalProperties requires this to be either a boolean or an object,
but it is currently defined as a list. This breaks validation for us.

> The additionalProperties keyword may be either a boolean or an object

https://json-schema.org/understanding-json-schema/reference/object.html#id2

And objects to me sound like they describe an associative array, or
dictionary:

> Objects are the mapping type in JSON. They map “keys” to “values”.

https://json-schema.org/understanding-json-schema/reference/object.html#object

So let's fix this by replacing the spurious list definition with an
actual object.

Signed-off-by: Martin Weinelt <hexa@darmstadt.ccc.de>