Mention processing: use \b to match word boundaries.
The `WORD_BOUNDARY` regex was missing some characters; a couple I've run
across are ' (e.g. as in "see foo's link") and > (e.g. as in here's
something you missed in the backlog: <foo>: blah blah).

Regex has a builtin \b to detect word boundaries; probably better to use
that.
https://developer.mozilla.org/en-US/docs/Web/JavaScript/Guide/Regular_Expressions/Assertions

Since \b is a zero-width match, we can also simplify some of the
processing and just use the whole-regex match.