Refactor/ovehaul AdminRoom command handling
This makes them obtain information (like the IRC server or clientList)
lazily rather than eagerly, making sure that we don't require a server
to be specified for stuff that doesn't need it (like !help).

Also makes the command handler interface almost unified, in case this
comes in handy later on for some reason.

Fixes GH-1431.