Add tests to ensure publicity syncing works for +s modes (#1698)
* PgDataStore: Fix syntax of array parameters

Arrays passed as query parameters become PostgreSQL arrays. However,
`foo IN $1` is not valid syntax, as `IN` can only be used with
subqueries (`foo IN (SELECT ...)`) and lists (`foo IN ($1, $2)`), which
require parentheses.

`foo IN ($1)` is interpreted as a list and the expression is compared
with the whole array, not checked against its members.

The correct syntax for checking if a value matches any member of an
array parameter is `foo = ANY($1)`, described in
https://www.postgresql.org/docs/15/functions-comparisons.html#id-1.5.8.30.16.

Fix `getIrcChannelsForRoomIds` and `getRoomsVisibility`. While we're at
it, replace `getMappingsForChannelByOrigin`'s workable but awkward
building of a parameter list.

Signed-off-by: Jan Alexander Steffens (heftig) <jan.steffens@gmail.com>

* Add tests for publicity-syncing

* Await publicity

* Refactor to fix publicity sync

* Refactor

* changelog

* Ensure PgDataStore returns private for missing entries

---------

Signed-off-by: Jan Alexander Steffens (heftig) <jan.steffens@gmail.com>
Co-authored-by: Jan Alexander Steffens (heftig) <jan.steffens@gmail.com>